Use shared rippledSource module in generateSFields.js
